ff_ffplay: merge: b1f78632c681f32c4901a85aa7ee72bf9ca6fece
ffplay: do not block audio thread on WIN32

The windows SDL audio driver plays the old data in the buffer in a loop if it
is not updated in time. So instead of waiting for data and blocking the the
audio thread, return silence if no data is available.

Should fix ticket #2289.

Reviewed-by: Michael Niedermayer <michael@niedermayer.cc>
Signed-off-by: Marton Balint <cus@passwd.hu>